+++ tests/unittests/cli/test_dependencies.py
@@ -49,7 +49,7 @@ def test_oldest_packages_pyproject_toml(tmpdir):
             '    "abc>=0.1",\n',
             "]\n",
         ])
-    replace_oldest_version(req_files=[str(req_file)])
+    replace_oldest_version(req_files=str(req_file))
     with open(req_file) as fp:
         lines = fp.readlines()
     assert lines == [
@@ -57,5 +57,27 @@ def test_oldest_packages_pyproject_toml(tmpdir):
         "dependencies = [\n",
         '    "fire>0.2",\n',
         '    "abc==0.1",\n',
+        "]\n",
+    ]
+
+
+def test_replace_packages_pyproject_toml(tmpdir):
+    req_file = tmpdir / "pyproject.toml"
+    with open(req_file, "w") as fp:
+        fp.writelines([
+            "[project]\n",
+            "dependencies = [\n",
+            '    "fire>0.2",\n',
+            '    "abc>=0.1",\n',
+            "]\n",
+        ])
+    replace_package_in_requirements(req_files=str(req_file), old_package="fire", new_package="water")
+    with open(req_file) as fp:
+        lines = fp.readlines()
+    assert lines == [
+        "[project]\n",
+        "dependencies = [\n",
+        '    "water>0.2",\n',
+        '    "abc>=0.1",\n',
         "]\n",
     ]