+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+---
+
+## [UnReleased] - 2025-08-DD
+
+### Changed
+
+- CLI: replace dependencies also in `pyproject.toml` ([#423](https://github.com/Lightning-AI/utilities/pull/423))
+
+
 ---
 
 ## [0.15.1] - 2025-08-04